Refactored and added a check whether the trust height isn't below or equal to 0

+++ restate-sync.sh
@@ -1,8 +1,17 @@
+# STATE-SYNC REFRESHER for Tendermint or CometBFT-based protocols.
+# https://github.com/zenodeapp/restate-sync
+# ZENODE (https://zenode.app)
+
+# This script will refresh the state sync of your node by setting the
+# configuration for state-sync to LATEST_HEIGHT-HEIGHT_INTERVAL.
+# This also makes a backup of the priv_validator_state.json file,
+# but it does wipe your entire /data folder!
+
 if [ -z "$1" ] || [ -z "$2" ]; then
     echo ""
     echo "Usage:   sh $0 <BINARY_NAME> <NODE_DIR> [HEIGHT_INTERVAL] [RPC_SERVER_1] [RPC_SERVER_2]"
     echo ""
-    echo "Example: sh refresh-state-sync.sh genesisd .genesis 1000 \"https://26657.genesisl1.org:443\""
+    echo "Example: sh $0 genesisd .genesis 1000 \"https://26657.genesisl1.org:443\""
     echo "         This will refresh the state sync using a trust height of LATEST_BLOCK - 1000 and"
     echo "         sets the RPC server addresses to https://26657.genesisl1.org:443"
     echo ""
@@ -15,15 +24,16 @@ fi
 
 BINARY_NAME=$1
 NODE_DIR=$2
+NODE_PATH=$HOME/$NODE_DIR
+CONFIG_PATH=$NODE_PATH/config
+DATA_PATH=$NODE_PATH/data
 HEIGHT_INTERVAL=${3:-2000}
 RPC_SERVER_1=$4
 RPC_SERVER_2=${5:-$RPC_SERVER_1}
-CONFIG_PATH=~/$NODE_DIR/config
-RPC_SERVER_PROVIDED=false
 
 # Check if the node directory exists
-if [ ! -d "$HOME/$NODE_DIR" ]; then
-    echo "The folder $HOME/$NODE_DIR does not exist."
+if [ ! -d "$NODE_PATH" ]; then
+    echo "The folder $NODE_PATH does not exist."
     exit 1
 fi
 
@@ -39,7 +49,7 @@ else
     RPC_SERVER_PROVIDED=true
 fi
 
-echo "WARNING: State-syncing will wipe the $HOME/$NODE_DIR/data folder (a backup of priv_validator_state.json will be made though)."
+echo "WARNING: State-syncing will wipe the $DATA_PATH folder (a backup of priv_validator_state.json will be made though)."
 echo "Service $BINARY_NAME will get halted using systemctl stop $BINARY_NAME, if this doesn't match your setup, make sure to halt the node yourself first!"
 echo ""
 read -p "Do you want to continue? (y/N): " ANSWER
@@ -50,22 +60,28 @@ if [ "$ANSWER" != "y" ]; then
     exit 1
 fi
 
+# Query block height data
+LATEST_HEIGHT=$(curl -s $RPC_SERVER_1/block | jq -r .result.block.header.height)
+TRUST_HEIGHT=$((LATEST_HEIGHT - $HEIGHT_INTERVAL))
+TRUST_HASH=$(curl -s "$RPC_SERVER_1/block?height=$BLOCK_HEIGHT" | jq -r .result.block_id.hash)
+
+if [ "$TRUST_HEIGHT" -le 0 ]; then
+  echo "Error: trust_height cannot be less than or equal to zero. Your [TRUST_HEIGHT] might be too large for the current state of the blockchain or there is something wrong with the RPC server(s)."
+  exit 1
+fi
+
 # Stop process
 systemctl stop $BINARY_NAME
 
 # Back up validator state
-cp ~/$NODE_DIR/data/priv_validator_state.json ~/$NODE_DIR/priv_validator_state.json.bak
-
-LATEST_HEIGHT=$(curl -s $RPC_SERVER_1/block | jq -r .result.block.header.height); \
-TRUST_HEIGHT=$((LATEST_HEIGHT - $HEIGHT_INTERVAL)); \
-TRUST_HASH=$(curl -s "$RPC_SERVER_1/block?height=$BLOCK_HEIGHT" | jq -r .result.block_id.hash)
+cp $DATA_PATH/priv_validator_state.json $NODE_PATH/priv_validator_state.json.bak
 
 sed -i '/^\[statesync\]/,/^enable = / s/enable = .*/enable = true/' $CONFIG_PATH/config.toml
 sed -i 's/trust_height = .*/trust_height = '$TRUST_HEIGHT'/' $CONFIG_PATH/config.toml
 sed -i 's/trust_hash = .*/trust_hash = "'"$TRUST_HASH"'"/' $CONFIG_PATH/config.toml
 
 # Only set the RPC server if we provided one
-if $RPC_SERVER_PROVIDED; then
+if [ "$RPC_SERVER_PROVIDED" = true ]; then
   sed -i 's#rpc_servers = .*#rpc_servers = "'"$RPC_SERVER_1,$RPC_SERVER_2"'"#' $CONFIG_PATH/config.toml
 fi
 
@@ -73,7 +89,7 @@ fi
 $BINARY_NAME tendermint unsafe-reset-all
 
 # Move backed up validator state back
-mv ~/$NODE_DIR/priv_validator_state.json.bak ~/$NODE_DIR/data/priv_validator_state.json
+mv $NODE_PATH/priv_validator_state.json.bak $DATA_PATH/priv_validator_state.json
 
 echo ""
 echo "New trust_height set to $TRUST_HEIGHT with trust_hash: $TRUST_HASH."
